check prior for skopt 0.9 version
