+++ ai4water/hyperopt/__init__.py
@@ -1,6 +1,6 @@
 
 """HyperOpt submodule"""
-from .utils import Real
-from .utils import Integer
-from .utils import Categorical
+from ._space import Real
+from ._space import Integer
+from ._space import Categorical
 from .hyper_opt import HyperOpt
